Refactor InternalNakadiException
