+++ src/main/java/org/zalando/nakadi/service/subscription/SubscriptionValidationService.java
@@ -9,10 +9,9 @@ import org.zalando.nakadi.domain.EventTypePartition;
 import org.zalando.nakadi.domain.NakadiCursor;
 import org.zalando.nakadi.domain.Subscription;
 import org.zalando.nakadi.domain.SubscriptionBase;
-import org.zalando.nakadi.exceptions.InternalNakadiException;
-import org.zalando.nakadi.exceptions.runtime.InvalidCursorException;
-import org.zalando.nakadi.exceptions.NakadiException;
 import org.zalando.nakadi.exceptions.runtime.InconsistentStateException;
+import org.zalando.nakadi.exceptions.runtime.InternalNakadiException;
+import org.zalando.nakadi.exceptions.runtime.InvalidCursorException;
 import org.zalando.nakadi.exceptions.runtime.NoEventTypeException;
 import org.zalando.nakadi.exceptions.runtime.RepositoryProblemException;
 import org.zalando.nakadi.exceptions.runtime.ServiceTemporarilyUnavailableException;
@@ -158,7 +157,7 @@ public class SubscriptionValidationService {
             }
         } catch (final InvalidCursorException ex) {
             throw new WrongInitialCursorsException(ex.getMessage(), ex);
-        } catch (final NakadiException | ServiceTemporarilyUnavailableException ex) {
+        } catch (final InternalNakadiException | ServiceTemporarilyUnavailableException ex) {
             throw new RepositoryProblemException("Topic repository problem occurred when validating cursors", ex);
         }
     }
