+++ src/main/java/org/zalando/nakadi/service/subscription/state/StartingState.java
@@ -8,6 +8,7 @@ import org.zalando.nakadi.domain.SubscriptionBase;
 import org.zalando.nakadi.domain.Timeline;
 import org.zalando.nakadi.exceptions.NakadiException;
 import org.zalando.nakadi.exceptions.NakadiRuntimeException;
+import org.zalando.nakadi.exceptions.runtime.InternalNakadiException;
 import org.zalando.nakadi.exceptions.runtime.NoStreamingSlotsAvailable;
 import org.zalando.nakadi.exceptions.runtime.AccessDeniedException;
 import org.zalando.nakadi.exceptions.runtime.ServiceTemporarilyUnavailableException;
@@ -158,7 +159,7 @@ public class StartingState extends State {
                         try {
                             // get oldest active timeline
                             return timelineService.getActiveTimelinesOrdered(et).get(0);
-                        } catch (final NakadiException e) {
+                        } catch (final InternalNakadiException e) {
                             throw new NakadiRuntimeException(e);
                         }
                     })
@@ -197,7 +198,7 @@ public class StartingState extends State {
                             // get newest active timeline
                             final List<Timeline> activeTimelines = timelineService.getActiveTimelinesOrdered(et);
                             return activeTimelines.get(activeTimelines.size() - 1);
-                        } catch (final NakadiException e) {
+                        } catch (final InternalNakadiException e) {
                             throw new NakadiRuntimeException(e);
                         }
                     })
