+++ src/test/java/org/zalando/nakadi/repository/db/CachingEventTypeRepositoryTest.java
@@ -3,7 +3,7 @@ package org.zalando.nakadi.repository.db;
 import org.junit.Test;
 import org.mockito.Mockito;
 import org.zalando.nakadi.domain.EventType;
-import org.zalando.nakadi.exceptions.InternalNakadiException;
+import org.zalando.nakadi.exceptions.runtime.InternalNakadiException;
 import org.zalando.nakadi.repository.EventTypeRepository;
 
 import static org.mockito.Mockito.mock;
