+++ src/test/java/org/zalando/nakadi/service/converter/VersionZeroConverterTest.java
@@ -2,7 +2,7 @@ package org.zalando.nakadi.service.converter;
 
 import org.junit.Before;
 import org.junit.Test;
-import org.zalando.nakadi.exceptions.InternalNakadiException;
+import org.zalando.nakadi.exceptions.runtime.InternalNakadiException;
 import org.zalando.nakadi.exceptions.runtime.InvalidCursorException;
 import org.zalando.nakadi.exceptions.runtime.NoSuchEventTypeException;
 import org.zalando.nakadi.exceptions.runtime.ServiceTemporarilyUnavailableException;
